requirements.txt
beautifulsoup4==4.7.1
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coverage==4.5.3
Flask==1.0.3
Flask-Login==0.4.1
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.0
flask-whooshee==0.7.0
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Markdown==3.1
MarkupSafe==1.1.1
micawber-bs4-classes==0.4.3
nose2==0.9.1
Pygments==2.4.2
python-dotenv==0.10.2
python-slugify==3.0.2
six==1.12.0
soupsieve==1.9.1
SQLAlchemy==1.3.3
text-unidecode==1.2
Werkzeug==0.15.4
Whoosh==2.7.4
WTForms==2.2.1
